+++ erppeek.py
@@ -7,12 +7,10 @@ Author: Florent Xicluna
 """
 import _ast
 import atexit
-import collections
 import csv
 import functools
 import optparse
 import os
-from pprint import pprint
 import re
 import shlex
 import sys
@@ -21,31 +19,16 @@ import traceback
 try:                    # Python 3
     import configparser
     from threading import current_thread
-    from xmlrpc.client import Fault, ServerProxy
-    basestring = str
-    int_types = int
-    _DictWriter = csv.DictWriter
+    from xmlrpc.client import Fault, ServerProxy, MININT, MAXINT
+    PY2 = False
 except ImportError:     # Python 2
     import ConfigParser as configparser
     from threading import currentThread as current_thread
-    from xmlrpclib import Fault, ServerProxy
-    int_types = int, long
-
-    class _DictWriter(csv.DictWriter):
-        """Unicode CSV Writer, which encodes output to UTF-8."""
-
-        def writeheader(self):
-            # Method 'writeheader' does not exist in Python 2.6
-            header = dict(zip(self.fieldnames, self.fieldnames))
-            self.writerow(header)
-
-        def _dict_to_list(self, rowdict):
-            rowlst = csv.DictWriter._dict_to_list(self, rowdict)
-            return [cell.encode('utf-8') if hasattr(cell, 'encode') else cell
-                    for cell in rowlst]
+    from xmlrpclib import Fault, ServerProxy, MININT, MAXINT
+    PY2 = True
 
 
-__version__ = '1.6.1'
+__version__ = '1.6.3'
 __all__ = ['Client', 'Model', 'Record', 'RecordList', 'Service',
            'format_exception', 'read_config', 'start_odoo_services']
 
@@ -78,7 +61,6 @@ Usage (some commands):
                                     # Upgrade the modules
 """
 
-STABLE_STATES = ('uninstallable', 'uninstalled', 'installed')
 DOMAIN_OPERATORS = frozenset('!|&')
 # Supported operators are:
 #   =, !=, >, >=, <, <=, like, ilike, in, not like, not ilike, not in,
@@ -113,6 +95,28 @@ _obsolete_methods = {
 }
 _cause_message = ("\nThe above exception was the direct cause "
                   "of the following exception:\n\n")
+_pending_state = ('state', 'not in',
+                  ['uninstallable', 'uninstalled', 'installed'])
+
+if PY2:
+    int_types = int, long
+
+    class _DictWriter(csv.DictWriter):
+        """Unicode CSV Writer, which encodes output to UTF-8."""
+
+        def writeheader(self):
+            # Method 'writeheader' does not exist in Python 2.6
+            header = dict(zip(self.fieldnames, self.fieldnames))
+            self.writerow(header)
+
+        def _dict_to_list(self, rowdict):
+            rowlst = csv.DictWriter._dict_to_list(self, rowdict)
+            return [cell.encode('utf-8') if hasattr(cell, 'encode') else cell
+                    for cell in rowlst]
+else:   # Python 3
+    basestring = str
+    int_types = int
+    _DictWriter = csv.DictWriter
 
 
 def _memoize(inst, attr, value, doc_values=None):
@@ -149,7 +153,10 @@ def literal_eval(expression, _octal_digits=frozenset('01234567')):
     node = compile(expression, '<unknown>', 'eval', _ast.PyCF_ONLY_AST)
     if expression[:1] == '0' and expression[1:2] in _octal_digits:
         raise SyntaxError('unsupported octal notation')
-    return _convert(node.body)
+    value = _convert(node.body)
+    if isinstance(value, int_types) and not MININT <= value <= MAXINT:
+         raise ValueError('overflow, int exceeds XML-RPC limits')
+    return value
 
 
 def is_list_of_dict(iterator):
@@ -341,19 +348,20 @@ class Service(object):
     _rpcpath = ''
     _methods = ()
 
-    def __init__(self, server, endpoint, methods, verbose=False):
+    def __init__(self, server, endpoint, methods,
+                 transport=None, verbose=False):
         if isinstance(server, basestring):
             self._rpcpath = rpcpath = server + '/xmlrpc/'
-            proxy = ServerProxy(rpcpath + endpoint, allow_none=True)
-            self._dispatch = proxy._ServerProxy__request
+            proxy = ServerProxy(rpcpath + endpoint,
+                                transport=transport, allow_none=True)
             if hasattr(proxy._ServerProxy__transport, 'close'):   # >= 2.7
                 self.close = proxy._ServerProxy__transport.close
+            rpc = proxy._ServerProxy__request
         elif server._api_v7:
-            proxy = server.netsvc.ExportService.getService(endpoint)
-            self._dispatch = proxy.dispatch
+            rpc = server.netsvc.ExportService.getService(endpoint).dispatch
         else:   # Odoo v8
-            self._dispatch = functools.partial(server.http.dispatch_rpc,
-                                               endpoint)
+            rpc = functools.partial(server.http.dispatch_rpc, endpoint)
+        self._dispatch = rpc
         self._endpoint = endpoint
         self._methods = methods
         self._verbose = verbose
@@ -414,20 +422,20 @@ class Client(object):
     _config_file = os.path.join(os.curdir, CONF_FILE)
 
     def __init__(self, server, db=None, user=None, password=None,
-                 verbose=False):
-        if isinstance(server, basestring) and server[-1:] == '/':
-            server = server.rstrip('/')
-        elif isinstance(server, list):
+                 transport=None, verbose=False):
+        if isinstance(server, list):
             appname = os.path.basename(__file__).rstrip('co')
             server = start_odoo_services(server, appname=appname)
+        elif isinstance(server, basestring) and server[-1:] == '/':
+            server = server.rstrip('/')
         self._server = server
-        float_version = 999
+        float_version = 99.0
 
         def get_proxy(name):
             methods = list(_methods[name]) if (name in _methods) else []
             if float_version < 8.0:
                 methods += _obsolete_methods.get(name) or ()
-            return Service(server, name, methods, verbose=verbose)
+            return Service(server, name, methods, transport, verbose=verbose)
         self.server_version = ver = get_proxy('db').server_version()
         self.major_version = re.match('\d+\.?\d*', ver).group()
         float_version = float(self.major_version)
@@ -650,6 +658,17 @@ class Client(object):
                                     user_password)
         return self.login('admin', user_password, database=database)
 
+    def duplicate_database(self, passwd, db_original_name, db_name,
+                           user_password='admin'):
+        """Duplicate an existing database.
+
+        The superadmin `passwd`, `db_original_name` and `db_name` are
+        mandatory.
+        Wait for the thread to finish and login if successful.
+        """
+        self.db.duplicate_database(passwd, db_original_name, db_name)
+        return self.login('admin', user_password, database=db_name)
+
     def execute(self, obj, method, *params, **kwargs):
         """Wrapper around ``object.execute`` RPC method.
 
@@ -748,17 +767,34 @@ class Client(object):
 
     def _upgrade(self, modules, button):
         # First, update the list of modules
-        updated, added = self.execute('ir.module.module', 'update_list')
+        ir_module = self.model('ir.module.module', False)
+        updated, added = ir_module.update_list()
         if added:
             print('%s module(s) added to the list' % added)
         # Find modules
-        ids = modules and self.search('ir.module.module',
-                                      [('name', 'in', modules)])
+        ids = modules and ir_module.search([('name', 'in', modules)])
         if ids:
-            # Click upgrade/install/uninstall button
-            self.execute('ir.module.module', button, ids)
-        mods = self.read('ir.module.module',
-                         [('state', 'not in', STABLE_STATES)], 'name state')
+            # Safety check
+            mods = ir_module.read([_pending_state], 'name state')
+            if mods:
+                raise Error('Pending actions:\n' + '\n'.join(
+                    ('  %(state)s\t%(name)s' % mod) for mod in mods))
+            if button == 'button_uninstall':
+                # Safety check
+                names = ir_module.read([('id', 'in', ids),
+                                        'state != installed'], 'name')
+                if names:
+                    raise Error('Not installed: %s' % ', '.join(names))
+                # A trick to uninstall dependent add-ons
+                ir_module.write(ids, {'state': 'to remove'})
+            try:
+                # Click upgrade/install/uninstall button
+                self.execute('ir.module.module', button, ids)
+            except Exception:
+                if button == 'button_uninstall':
+                    ir_module.write(ids, {'state': 'installed'})
+                raise
+        mods = ir_module.read([_pending_state], 'name state')
         if not mods:
             if ids:
                 print('Already up-to-date: %s' %
@@ -924,10 +960,12 @@ class Client(object):
             domain.append(('state', op, ['uninstalled', 'uninstallable']))
         mods = self.read('ir.module.module', domain, 'name state')
         if mods:
-            res = collections.defaultdict(list)
+            res = {}
             for mod in mods:
+                if mod['state'] not in res:
+                    res[mod['state']] = []
                 res[mod['state']].append(mod['name'])
-            return dict(res)
+            return res
 
     def keys(self, obj):
         """Wrapper for :meth:`Model.keys` method."""
@@ -1312,6 +1350,10 @@ class RecordList(object):
             msg = "has no attribute %r"
         raise AttributeError("'RecordList' object %s" % msg % attr)
 
+    def __eq__(self, other):
+        return (isinstance(other, RecordList) and
+                self.id == other.id and self._model is other._model)
+
 
 class Record(object):
     """A class for all Odoo records.
@@ -1347,6 +1389,12 @@ class Record(object):
 
     def __str__(self):
         return self._name
+
+    if PY2:
+        __unicode__ = __str__
+
+        def __str__(self):
+            return self._name.encode('ascii', 'backslashreplace')
 
     def _get_name(self):
         try:
@@ -1507,16 +1555,18 @@ class Record(object):
 
 def _interact(global_vars, use_pprint=True, usage=USAGE):
     import code
-    try:
-        import builtins
-        _exec = getattr(builtins, 'exec')
-    except ImportError:
-        def _exec(code, g):
-            exec('exec code in g')
+    import pprint
+    if PY2:
         import __builtin__ as builtins
 
+        def _exec(code, g):
+            exec('exec code in g')
+    else:
+        import builtins
+        _exec = getattr(builtins, 'exec')
+
     if use_pprint:
-        def displayhook(value, _printer=pprint, _builtins=builtins):
+        def displayhook(value, _printer=pprint.pprint, _builtins=builtins):
             # Pretty-format the output
             if value is None:
                 return