+++ CHANGES.rst
@@ -2,11 +2,29 @@ Changelog
 ---------
 
 
-1.x (unreleased)
-~~~~~~~~~~~~~~~~
+1.6.3 (2015-12-30)
+~~~~~~~~~~~~~~~~~~
+
+* Do not parse long integers which overflow in XML-RPC.
+
+
+1.6.2 (2015-09-17)
+~~~~~~~~~~~~~~~~~~
+
+* Add an optional ``transport`` argument to the ``Client`` constructor.
+  This is useful for tweaking the SSL context or adding an optional
+  timeout parameter.
+
+* Implement ``==`` comparison for ``RecordList`` instances.
+
+* Uninstall dependent add-ons in a single call.
+
+* Do not install/uninstall add-ons if other actions are pending.
 
 * Do not hang when the ``Client`` constructor receives invalid
   arguments.
+
+* Fix ``str(record)`` and ``print(record)`` with non-ASCII names.
 
 
 1.6.1 (2014-11-12)